Backend/requirements.txt
# Django + DRF
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9

# Task Queue
celery[redis]>=5.3
redis>=5.0

# Environment
python-dotenv>=1.0

# Encryption
cryptography>=42.0

# ML Dependencies (for bridge services — models are NOT retrained)
joblib>=1.3
pandas>=2.1
numpy>=1.26
xgboost>=2.0
scikit-learn>=1.4

# File parsing
openpyxl>=3.1

# Production server
gunicorn>=22.0
